[Pallas:MGPU] Make semaphores compatible with plgpu.kernel and the profiler
The previous code was overly specific to pl.pallas_call and did not work with
plgpu.kernel at all. Now, semaphores can be allocated using run_scoped, which also
has the interesting side effect of the allocations being collective within
each program. For a persistent kernel that means that a program/block can communicate
with programs/blocks on other devices that have the same program ID, but it is
currently impossible to e.g. synchronize all programs in a grid. We don't have a
use case for it now, so we can add it later.

PiperOrigin-RevId: 761006798